Initial SPP
Initial Soft Patch Panel code

SPP is a framework for managing DPDK resources.

The goal of SPP is to easily interconnect DPDK applications together,
and assign resources dynamically to these applications to build a
pipeline.

The first version of SPP provides for the management of DPDK ports, and
assigning ports to different DPDK applications.

The framework is composed of a primary DPDK application that is
responsible for resource management. This primary application doesn't
interact with any traffic, and is used to manage creation and freeing of
resources only.

A Python based management interface is provided to control the primary
DPDK application to create resources, which are then to be used by
secondary applications. This management application provides a socket
based interface for the primary and secondary DPDK applications to
interface to the manager. The management application will use OVSDB to
maintain all created and assigned ports.

Signed-off-by: Ferruh Yigit <ferruh.yigit@intel.com>
Signed-off-by: Gerald Rogers <gerald.rogers@intel.com>
Signed-off-by: Sy Jong Choi <sy.jong.choi@intel.com>