doc: correct assignment of secondary ID
Although assignment of secondary ID is started from 0 in setup guide,
it is started from 1 actually and only assigns 0 if given ID is used
already.

This change is for correcting assignment of secondary ID from 1 in
command line examples and diagrams.

It also includes update for misc bug fixes of markdown.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>