spp: fix bug of precmd
In precmd function, command log is not recorded if it is
'playback' to avoid illegal action when the log is loaded.
However, 'bye' should be also excluded from the log because
spp.py is terminated immediately if the log is loaded.
This change for adding a condition for command logging.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>