spp/shared: refactor printing port status
Port status of each of clients is printed with print_active_ports().
It is defined in both of spp_nfv and spp_vm. It is redundant and should
be changed to be a common library function.

This update is to move print_active_ports() in shared/common.c as common
library. This update also following changes.
* client_id, ports_fwd_array and port_map are added as arguments
  of this function to be enable them referred.
* Print 'none' instead of port_id if port is unassigned.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>