spp: declare global variables explicitly
In spp.py, connections are managed as global variables and used from
functions without declaring as global. It might cause an error because
updating variable does not affect outside of the scope if it is not
declared as global.

This patch is for declaring global explicitly. It also includes
following updates.

* Remove arguments for passing MAIN2SEC or other global variables
  to __init__() to avoid misunderstanding for local variables.

* Move check_sec_cmds() into Shell class because only used in it.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>