spp_nfv: enable to patch ports with resource ID
In spp, port ID is assigned as a number incrementally and the number of
ID can be different from each sec processes. It depends on the order of
adding ports.

It is confusing to users for patching because each of sec processes has
different port IDs but they refer the same resource. It might cause a
mistake.

This update enables users to patch ports with a combination of port type
and number to identify resources as following.
  spp> sec 1;patch phy:0 ring:0

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>