spp: move controller to sub directory
SPP controller 'spp.py' is monolithic and has got to be large. It is
hard to be maintained because several classes are included in the file
and global variables are shared among its instances.

To improve maintainability, 'spp.py' is divided into main and class
files. Global variables are also moved to shared file 'spp_common.py'.

This update is halfway for refactoring and classes are going to be
more fine-grained.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>