controller: add do_topo to shell.py
'topo' is an experimental command for SPP controller to outptu topology
of network configuration. It enables to display a graphical image of
network configuration on terminal, or browser for which websocket
server is running with.

'topo' command supports four types of output. For multiple output
support, it uses graphviz, imagemagick and other external tools.
* terminal (but very few terminals supporting to display images)
* browser (websocket server is required)
* image file (jpg, png, bmp)
* text (dot, json, yaml)

Here is a list of required tools for 'topo'. Notice that SPP controller
and 'topo' command support MacOS on a remote host.
* graphviz
* imagemagick
* libsixel-bin (for Ubuntu) and terminal app supporting img2sixel
* iTerm2 and imgcat (for MacOS)

For terminal, only few terminal applications are supported. You can use
mlterm or xterm for which 'img2sixel' supports on linux
or iTerm on MacOS. If you use iTerm2, you have to get a shell script
'imgcat' from [1] and save it as
'spp/src/controller/3rd_party/imgcat.sh'.

[1] https://iterm2.com/documentation-images.html

For browser support, you are required to install websocket libraries and
use 'spp_ws' tool. 'spp_ws' is not released yet, but soon.

There are some example usecases.

(1) Display on terminal or browser.
spp > topo term
spp > topo http

(2) Output to a file.
spp > topo myconfig.jpg
spp > topo myconfig dot

This update is only for 'shell.py' and requires a module file 'topo.py'
in which methods for 'topo' command are implemented.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>