controller: add topo_subgraph command
'topo_subgraph' is for defining subgraph of resources in a network
topology generated by 'topo'. Typically it is used for grouping vhost
interfaces of each of VMs or containers to be more understandable the
generated network topology.

Usage: topo_subgraph [VERB] [LABEL] [SUBGRAPH]
       VERB: 'add' or 'del'

If you assign vhost1 and vhost2 to VM1, define subgraph as following.

  spp > topo_subgraph add VM1 vhost1,vhost2

Or delete defined subgraph.

  spp > topo_subgraph del VM1

To show all of defined subgraphs, simply run without arguments.

  spp > topo_subgraph

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>