controller: change displaying status
This update is for removing port_id from the result of status
command because bahaviour of port_id assignment is changed from
DPDK 18.02 and it possibly confuses users.

Secondary's status message is changed from

  spp > sec 1;status
  port_id:0,on,PHY,outport:none
  port_id:1,on,PHY,outport:none
  port_id:2,on,RING(0),outport:none

to be more simplified YAML like style.

  spp > sec 1;status
  status: idling
  ports:
    - 'phy:0 -> ring:0'
    - 'phy:1'
    - 'ring:0 -> phy:1'

SPP controller expects message format from secondary is as following to
decode it to show as above example.

  status: running\nports: phy:0-vhost:0,vhost:0-null

This update is only for controller. Updates for secondary processes are
included in next patches.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>