controller: add delimiter for topo command
Resource ID consists of resource type and ID separated with
delimiter ':', for instance, 'phy:0'. This delimiter causes a parse
error in topo command because graphviz does not support ':' as a valid
character for node ID. If ':' is included in node ID, graphviz outputs
an imcomplete graph.

To avoid this problem, replace ':' with other valid character for node
IDs while labels displayed in the graph are not replaced.

This update add definitions for charactors of delimiter and label in
'spp_common.py'. Delimiter ':' is replaced with label charactor '_'
while graphviz script is generated.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>