spp_nfv: update for creating rx pcap device
Libpcap-based PMD of DPDK requires two different streams for rx and tx
of pcap files[1]. 'tx_pcap' is created by DPDK but not for 'rx_pcap'.
In SPP, rx pcap file should be prepared like as '/tmp/spp-rx1.pcap'
before 'add pcap', or failed to setup pcap PMD.

This update is for creating the rx pcap with 'text2pcap'[2] if it does
not exist.

This update is also including definition of format of pcap files in
src/shared/common.h as following.

  #define PCAP_IFACE_RX "/tmp/spp-rx%d.pcap"
  #define PCAP_IFACE_TX "/tmp/spp-tx%d.pcap"

[1] https://dpdk.org/doc/guides/nics/pcap_ring.html
[2] https://www.wireshark.org/docs/man-pages/text2pcap.html

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>