controller: convert license of Intel to SPDX tag
To change license description of SPP controller, it must be considered
that it was originally just one file, but divided into several files
and changed to improve maintainalibity and easily extensible.

Here is a list of all files of SPP controller. (1) is provided by Intel
and (2) is NTT.
* src/controller/command/hello.py  (2)
* src/controller/conn_thread.py    (1)
* src/controller/shell.py          (1)
* src/controller/shell_lib/common.py  (2)
* src/controller/spp.py            (1)
* src/controller/spp_common.py     (1)
* src/controller/topo.py           (2)
* src/spp.py                       (1)

Updating license is divided in two steps. First, update for Intel and
then NTT. This is the first patch.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>
Acked-by: Ferruh Yigit <ferruh.yigit@intel.com>