controller: change importing for python3
This is the first patch for updating to python3. To keep backward
compatibility with python2, it uses future module.

To comply with python3, it uses absolute_import module in __future__
and changes importing modules with path info. Some of module name, for
instance Queue or SocketServer, are also changed to the names of
python3's.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>