controller: fix bug of topo to refer external cmd
'topo term' command uses img2sixel (or imgcat for iTerm2) for generating
an image of topology, but there is no checking if the command exists.

This update is to add checking it before generating image. If the
command is not existing, 'topo term' shows an error message and do
nothing.

This update is also including an change of the name of imgcat script
from 'imgcat.sh' to 'imgcat'. It is just to simplify setting up.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>