recipes: add recipes directory
This patch is to introduce recipes. Recipe is a set of instruction
to configure SPP network. This configuration is a set of SPP command
actually.

Recipe is loaded with 'playback' command to restore the network
configuration. This is an example.

    spp > playback recipes/spp/1nfv_bi.config

Some of recipes are already included in docs/, but moved to recipes/
by this update.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>