docs: add index of SPP container tool
SPP container is a management tool for running DPDK application
on a container. SPP container provides a set of python scripts for
buiding docker image and launching app containers.Containers are
connected with SPP and its network is to be easily configurable.

This patch is for adding an index file of documents of SPP container.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>