spp_vf: change to not use flush and cancel commnad
It is bothering to run flush every time after changing configrations.
User does not need to run flush if spp_vf executes it for accepting
command by itself. Cancel command is also not needed for this change.

This patch is to stop support for flush and cancel command by following
changes.
- Send an error if it receives flush or cancel command.
- Call spp_flush() if it receives component, port or classifier_mac
  command.
- Remove functions for cancel command.

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>
Acked-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>