spp_vf: fix unexpected tx port is shown in status
Unexpected tx port is shown without adding it because of bug of status
command. Here is an example.

  spp > sec 1;component start fwd1 12 forward
  spp > sec 1;port add ring:0 rx fwd1
  spp > sec 1;flush
  spp > sec 1;status
  { "results": [ { "result": "success" } ],"info": { "client-id": 2,
  ...
  "tx_port": [ { "port": "phy:0", "vlan": { "operation": # Here
  ...

It is because status command wrongly assumes the number of tx port is
equal to rx if the number of rx port is greater than zero. To fix the
problem, add 'num_tx' in struct forward_path to hold the number exactly.

Signed-off-by: Hideyuki Yamashita <yamashita.hideyuki@po.ntt-tx.co.jp>
Signed-off-by: Naoki Takada <takada.naoki@lab.ntt.co.jp>
Acked-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>