project: update gitignore to exclude PDF files
To compile a PDF document, the format of images included in the
document should also be PDF. It means that SVG images should be
converted to PDF while the PDF document is generated and not be
managed from git.

To exclude temporarily generated PDF images, add an entry to
'.gitignore'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>