tools/sppc: build image with custom git repo
SPP container supports to create an image with a custom repository other
than on dpdk.org. However, the name of directory should be the same as
them. For example, the name of custom DPDK repo should also be 'dpdk'.

This patch is for supporting to use any of directory name. The name of
directory is extracted from the git URL. For example, directory name of
'https://github.com/user/custom-dpdk.git' is 'custom-dpdk'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>