tools/sppc: add paths to Dockerfiles of SPP
If you clone a custom SPP repo and its directory name is not 'spp',
you failed to launch an appcontainer script because the scripts does
not know the name of custom repo and cannot find SPP's executable file.

To be found the executable in container, this patch adds PATH entries
of the executables of custom repo to Dockerfile. By this change.
appcontainer can find the executables without the custom repo's name
while launching appcontainer.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>