tools/sppc: add workdir docker option
Some of apps require to be launched from specific directory even if PATH
is configured properly. For instance, pktgen should be launched from
project root because for finding `Pktgen.lua` in the directory.

This update is for adding workdir option for specifying a directory
from which the app is launched.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>