tools/sppc: add paths to Dockerfiles of pktgen
If you clone a custom pktgenrepo and its directory name is not
'pktgen-dpdk',
you failed to launch an appcontainer script because the scripts does
not know the name of custom repo and cannot find executable.

To be found the executable in container, this patch adds PATH entries
of the executables of custom repo to Dockerfile. By this change.
appcontainer can find the executables without the custom repo's name
while launching appcontainer.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>