tools/sppc: update launcer for custom pktgen repo
To launch executable in custom repo, change the path from absolute
path to just filename. Path of the file is defined as PATH in each
of Dockerfiles and no need to use absolute path by previous change.
In addition, pktgen should be launched from the directory which includes
`Pktgen.lua` but cannot found it if you use custom repo. To avoid it,
add specifying workdir for launching from the directory.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>