controller: add history and redo commands
SPP controller has no history for commands. It is useful if user inputs
several commands and check it step by step. It is also useful to redo
the commands by referring the history.

This update is to add history and redo commands. History command shows
the list of command with index number. Command history is saved as
'.spp_history' which is created in the project root directory.

  spp > history
    1  sec 1;status
    2  pri;status
    ...

User can redo a previous command with index number.

  spp > redo 1  # run 'sec 1;status' from redo command
  - status: running
  - port:
    - phy:0 -> ring:0

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>