shared: update status format of spp_nfv
Update message format of get_sec_status_json() for considering spp-ctl's
API. The message should be the same as the response params of spp-ctl
for maintainance. 'client_id' is added to the args for including it to
the message.

This update also changes spp_nfv and spp_vm for adding 'client_id'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>