controller: update parsing status message
To apply with the update of spp_nfv and spp_vm, update parsing the
message of status.

In the updated message, 'ports' attribute is changed to not include
patch info and it is moved to the new 'patches' attribute. This patch
is to parse 'patches' attribute. Output of status command is not
changed.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>