controller: change controller to use spp-ctl
SPP controller manages primary and secondary processes, and create TCP
sessions while launching. Other process is not allowed to manage SPP
processes directly without using management port. It is difficult to
manage from other process via management port for supporting detailed
operations.

It is better to manage SPP by spp-ctl and to change SPP controller as a
client because it is allowed to maange it from multiple controllers via
well-defined REST APIs. This update is to change SPP controller to a
client of spp-ctl.

* Remove connection thread classes and queues because SPP controller is
  no need to has sessions.

* Discard management port support because spp-ctl plays the role.

* Use SppCtlClient for requesting to spp-ctl.

* Change parsers for responses from primary and secondary processes to
  spp-ctl.

* Remove arguments IP address and ports of primary and secondary
  processes.

* Add SppCtlClient class for requests and responses.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>