controller: move pri command to SppPrimary
'spp.py' includes many methods and vals for supporting several commands.
Some of them are global and others are command's local. It would be hard
to maintain if more commands are added to 'spp.py'.

SppPrimary defines 'pri' command and its completion as in a separated
module. It is intended to be used from Shell, which is derived from
'cmd.Cmd'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>