controller: add SppVf class
Spp_vf has dedicated controller 'spp_vf.py' and is not managed from
'spp.py'. It is confusing to users.

Thanks to spp-ctl, it is easy to send a request to spp_vf from 'spp.py'
without any effort. This update is to add SppVf class behaviour as a
client for spp-ctl. An instance of the class is intended to be used
from do_vf() and complete_vf() methods of Shell class in 'spp.py'.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>