controller: add vf command to SPP controller
To manage spp_vf from 'spp.py', add 'vf' command to Shell class. Spp_vf
might have several instances as similar to spp_nfv, and deciding which
of ones is also similar to. 'vf' command consists of a indicator and
actual command. Here is an example.

  spp > vf 3; component start fw1 5 forward

In this example, indicator 'vf 3;' is before spp_vf's command 'component
start ...'. The number in indicator is a secondary ID actually, so you
cannot assign the same ID of others.

You can refer the usage of 'vf' command with 'help' command.

  spp > help vf

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>