spp-ctl: change to bind IP address to SPP procs
Spp-ctl creates a socket with address `localhost` for SPP processes.
However, it should be able to change for some cases. For instance, SPP
container process requires to access spp-ctl running on host. It cannot
find spp-ctl from inside of container if the address is `localhost`.
This update is to change binding IP address to be found.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>