controller: add mirror command to SPP controller
This update is to 'mirror' command to Shell class for managing
spp_mirror from 'spp.py'. Spp_mirror might have several instances, and
deciding which of ones is also similar to other secondary processes.
'mirror' command consists of an indicator and actual command. Here is
an example.

  spp > mirror 3; component start mr1 5 mirror

In this example, the indicator 'mirror 3;' is before the command
'component start ...'. The number in indicator is a secondary ID
actually, so you cannot assign the same ID of others.

You can refer the usage of 'mirror' command with 'help' command.

  spp > help mirror

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>