docs: fix version in guides
When building HTML guides, the type of the version string was wrong:

WARNING: The config value `version' has type `bytes', defaults to `str'.
WARNING: The config value `release' has type `bytes', defaults to `str'.
[...]
Exception occurred:
  File "/usr/lib/python3.7/site-packages/sphinx/builders/html.py", line 479, in prepare_writing
    release = return_codes_re.sub('', self.config.release),
TypeError: cannot use a string pattern on a bytes-like object

The version number is also updated to match the latest tag.

The option --no-print-directory is added to make showversion command,
so only the version number is printed in the doc.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Acked-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>