controller: add multi-node support
This update is to manage multiple nodes of SPP from SPP CLI.

To add several nodes, give several address of spp-ctl with `-b` option.
Default address is `127.0.0.1:7777` and it is able to be overwritten.
If port number is not given, default value `7777` is remained.

  # Use two spp-ctl running on "192.168.1.101:7777" and
  # "192.168.1.102:7777".
  $ python src/spp.py -b 192.168.1.101 -b 192.168.1.102

`server` command is provided for switching spp-ctl under the management
of SPP CLI.

  # Show all of spp-ctl nodes.
  spp > server  # or 'server list'
    1: 192.168.1.101:7777 *
    2: 192.168.1.102:7777

  # Switch to second node.
  spp > server 2
  Switch spp-ctl to "2: 192.168.1.102:7777".

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>