docs: move design section to doc root
Design section is a subsection of overview section. However, it should
be moved to doc root if the number of SPP processes is increased.
In general, it is not good to have much subsections for getting
information.

This update is to move design section from overview to the doc root.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>