spp-ctl: fix error in parsing replied message
Spp-ctl expects a response of command from secondary process is JSON
format for checking the result, but it is not a constraint. However,
parsing is failed and terminated if it is not JSON because of
inappropriate error handling. This update is to add a exception handler
of JSONDecodeError and not terminated.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>