tools: add customized cpu_layout tool
`cpu_layout.py` is provided as a usertool of DPDK which is used to show
layout of sockets, cores and CPUs. This update is to add a customized
version of this tool to show the layout in JSON format. It is intended
to be used from other programs.

To output in JSON format, simply add `--json` option.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>