spp_primary: refactor var names for ring ports
Spp_priamry has ring ports and its stats as structure `client` and
`num_clients` because ring was used only for communication between
primary and secondary. However, it is used for packet forwarding
betweeen secondaries, not primary.

This update is to change the var names of `client` to `ring` to avoid
misunderstanding.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>