controller: add launch sub command in pri
As launch is supported in spp_primary and spp-ctl, add `launch` in
`pri` command. It takes type of secondary process, ID and arguments
of launching the process. Here is an example.

  spp > pri; launch nfv 1 -l 1,2 -m 512 -- -n 1 -s 192.168.1.100:7777

You notice that secondary ID is used two times. The reason is it is used
for completion. If you specify secondary ID, it completes typical
arguments. You just correct a few params of command line.

  spp > pri; launch  # press TAB
  mirror  nfv     pcap     nfv
  spp > pri; launch nfv  # press TAB
  1 10 11 12 2 3 4 ...
  spp > pri; launch nfv 1  # press TAB and correct some params
  spp > pri; launch nfv 1 -l 1,2 -m 512 -- -n 1 -s 192.168.1.100:7777

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>