controller: fix bug of path completion
SPP CLI uses `shell_lib.compl_common()` to completion path for commands
referring a file or a directory, but does not work correctly if the path
includes '-'. It is because Cmd unexpectedly treats '-' as a delimiter
as same as '/'. This patch is to fix this issue by correcting path
including '-' while serching candidates.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>