docs: divide spp-ctl API reference
As increased the number of process types, the number of lines of API
reference is getting large and hard to maintain. It is the time to
divided into several files.

This update is to divide the reference into sections of each of
processes. It also includes revising for updating old examples,
correcting typos, or adding descriptions for which lacks of explaining.

Here is the updated list of contents for which each of divided files is
corresponds to, but the list itself is the same as before version.

  1. spp-ctl REST API
  2. API Independent of Process Type
  3. API for spp_primary
  4. API for spp_nfv
  5. API for spp_vf
  6. API for spp_mirror
  7. API for spp_pcap

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>