controller: add config command
`config` command is for managing default configurations of SPP CLI.
For example, you can change default command prompt `spp > ` to another
one.

List of all of configurations is referred with `config`, or list of
started with `p` is referred with `config p`.

  # show all configs
  spp > config
  - topo_size: "60%"	# Percentage or ratio of topo
  - prompt: "spp > "	# Command prompt
  ...

  # only started with `p`
  spp > config p
  - prompt: "spp > "	# Command prompt

To change configuration, specify key and its value.

  # set prompt
  Set prompt: "$ spp "
  $ spp

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>