controller: change nof worker lcores configurable
This update is to change the number of worker lcores of each of
secondary processes configurable.

It is decides with secondary ID and config value referred from `config`
command.
  - sec_vf_nof_lcores
  - sec_mirror_nof_lcores
  - sec_pcap_nof_lcores

If secondary ID is `3` and `sec_vf_nof_lcores` is 3, woker lcores is
suggested as `3-5`, started from `3` and use three cores.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>