bin: add SPP launcher scripts
SPP is mult-process application, so primary process should be launched
before other processes. In addition, spp-ctl should be launched before
primary. Users should keep in mind the order of launching processes,
but make a mistake sometimes.

This update is to add launcher scripts in `bin/` for SPP. Users can
configure options defined in `bin/env.sh` and launch spp-ctl,
spp_primary and SPP CLI from `bin/start.sh`.

    $ bin/start.sh

In general, spp_primary has a port while launching and is failed to
launch if there are no ports. In this case, you need to add a vdev
and it supports `PRI_VHOST_IDS` in `bin/env.sh` which defines a list
of IDS of vhost devices attached to spp_primary. You should keep in mind
the total number of ports is the same as ports you have as
`PRI_PORTMASK`.

You can find log files of all of processes launched from this
`start.sh` in `spp/log` directory.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>