shared/sec: refactor parse error code
Revise the name of enum `spp_comamnd_decode_error_code` to
`sppwk_parse_error_code`, and its members started from `SPP_CMD_DERR_`
to `SPPWK_PARSE_` to be more reasonable. The prefix of the members does
not need to include the word `ERR` because following word, such as
`INVALID_VALUE`, can be understood as error.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>