shared/sec: revice cmd parser of SPP worker
The name of header file of command parser for SPP worker threads is
`command_dec.h` which means "decode commands", but the features are
not for decoding but parsing. This update is to change it to
`cmd_parser.h` and refactor name of vars, functions and its comments
defined in the header file.

This update also includes refactoring name of attributes of
sppwk_parse_err_msg object and comments for functions in which this
object is used.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>