shared/sec: revise port info of SPP worker
SPP worker has two kind of struct `spp_port_index` and `spp_port_info`,
but it is confusing and not clear for considering usages.
`spp_port_index` is used for the case just refering port type and index,
and `spp_port_info` is for more detailed info actually.

This update is to refactor for rename prefix `spp_` to `sppwk_`, and
adding comment for usages.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>