shared/sec: rename func for getting port ID
Function `spp_get_dpdk_port()` is for getting consistent port ID
provided by rte ethdev library, and intended to be used to convert SPP's
resource UID such as `phy:0` to consistent ID. The name of function is
not describing the purpose exactly. This update is to rename it to
`sppwk_get_ethdev_port_id()` for refactoring.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>