shared/sec: remove no meaning str defines
Defines of fixed string, such as `#define SPP_COMMAND_EXIT "exit"`, are
no meaning because this long define is referred just one time. In
addition, there are other similar defines which are actually required.
It makes code complex and hard to maintain. This update is to remove
such kind of defines.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>