shared/sec: fix funcs for getting int and uint
The name `get_uint_value()` was renamed as `get_vlan_uint_val()`, but
it is used lcore ID. `get_int_value()` was also renamed as
`get_vlan_int_val()` which is used for getting VLAN ID, but can be used
for other purposes. The term `vlan` should be removed from these funcs.
In addition, the names do not contain that checking the given value is
in the given range. This update is to fix the issues.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>