shared/sec: rename fname of command_proc
The name of file `command_proc.c` and its header has the meaning
"to process command sent from controller", but the term `process` is
is not clear in meaning in this context because noun is usually used as
file name and `process` means a computing entity.

This update is to rename to be clear as `cmd_runner.c`, and
also its header file.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>