shared/sec: rename func for getting component ID
The term `component ID` is lcore ID used by a SPP worker thread
actually. Fuctions for managing component ID should be renamed for
lcore ID because it is confusing to use several terms for the same
resource. This update is to rename the functions and variables of
component ID to lcore ID to be explicit in meaning.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>